Reorg of the EditPlantVC extensions
