+++ WaterMyPlants/WaterMyPlants/Extensions/EditPlantVC.swift
@@ -1,0 +1,67 @@
+//
+//  EditPlantVC.swift
+//  WaterMyPlants
+//
+//  Created by Craig Swanson on 3/9/20.
+//  Copyright © 2020 craigswanson. All rights reserved.
+//
+
+import Foundation
+import UIKit
+import Photos
+
+extension EditPlantViewController: UIImagePickerControllerDelegate, UINavigationControllerDelegate, UITextFieldDelegate {
+    
+    // MARK: - Methods
+    // Image picker controller
+    func presentImagePickerController() {
+        guard UIImagePickerController.isSourceTypeAvailable(.photoLibrary) else { return }
+        
+        let imagePicker = UIImagePickerController()
+        
+        imagePicker.sourceType = .photoLibrary
+        imagePicker.delegate = self
+        
+        present(imagePicker, animated: true, completion: nil)
+    }
+        
+        // MARK: Image Picker Delegate Method
+        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
+            guard let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage else { return }
+            plantImageView.image = resizeImage(image: image, targetSize: CGSize(width: 500.0, height: 500.0))
+            
+            picker.dismiss(animated: true, completion: nil)
+        }
+        
+        // MARK: - Image Resize
+        func resizeImage(image: UIImage, targetSize: CGSize) -> UIImage {
+            let size = image.size
+
+            let widthRatio  = targetSize.width  / size.width
+            let heightRatio = targetSize.height / size.height
+
+            // Figure out what our orientation is, and use that to form the rectangle
+            var newSize: CGSize
+            if widthRatio > heightRatio {
+                newSize = CGSize(width: size.width * heightRatio, height: size.height * heightRatio)
+            } else {
+                newSize = CGSize(width: size.width * widthRatio, height: size.height * widthRatio)
+            }
+
+            // This is the rect that we've calculated out and this is what is actually used below
+            let rect = CGRect(x: 0, y: 0, width: newSize.width, height: newSize.height)
+
+            // Actually do the resizing to the rect using the ImageContext stuff
+            UIGraphicsBeginImageContextWithOptions(newSize, false, 1.0)
+            image.draw(in: rect)
+            let newImage = UIGraphicsGetImageFromCurrentImageContext()
+            UIGraphicsEndImageContext()
+
+            return newImage!
+        }
+
+        func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+            self.view.endEditing(true)
+            return false
+        }
+}
