+++ WaterMyPlants/WaterMyPlants/View Controllers/EditPlantViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import Photos
 
-class EditPlantViewController: UIViewController, UIImagePickerControllerDelegate, UINavigationControllerDelegate, UITextFieldDelegate {
+class EditPlantViewController: UIViewController {
     
     // MARK: - Properties
     var plantController: PlantController?
@@ -78,6 +78,7 @@ class EditPlantViewController: UIViewController, UIImagePickerControllerDelegate
         present(imagePickerController, animated: true, completion: nil)
     }
     
+    // MARK: - Save Changes
     @IBAction func saveChanges(_ sender: UIBarButtonItem) {
         guard let plantController = plantController,
             let nickname = plantNickname.text,
@@ -136,20 +137,7 @@ class EditPlantViewController: UIViewController, UIImagePickerControllerDelegate
     @IBAction func cancel(_ sender: UIBarButtonItem) {
         dismiss(animated: true, completion: nil)
     }
-    
-    // MARK: - Methods
-    // Image picker controller
-    private func presentImagePickerController() {
-        guard UIImagePickerController.isSourceTypeAvailable(.photoLibrary) else { return }
-        
-        let imagePicker = UIImagePickerController()
-        
-        imagePicker.sourceType = .photoLibrary
-        imagePicker.delegate = self
-        
-        present(imagePicker, animated: true, completion: nil)
-    }
-    
+
     // Update views puts the appropriate values in the elements if a plant is passed in, otherwise empty text fields await the user
     private func updateValues() {
         guard isViewLoaded else { return }
@@ -167,47 +155,5 @@ class EditPlantViewController: UIViewController, UIImagePickerControllerDelegate
         } else {
             title = "Enter New Plant"
         }
-    }
-    
-    // MARK: Image Picker Delegate Method
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-        guard let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage else { return }
-        plantImageView.image = resizeImage(image: image, targetSize: CGSize(width: 500.0, height: 500.0))
-        
-        picker.dismiss(animated: true, completion: nil)
-    }
-    
-    // MARK: - Image Resize
-    func resizeImage(image: UIImage, targetSize: CGSize) -> UIImage {
-        let size = image.size
-
-        let widthRatio  = targetSize.width  / size.width
-        let heightRatio = targetSize.height / size.height
-
-        // Figure out what our orientation is, and use that to form the rectangle
-        var newSize: CGSize
-        if widthRatio > heightRatio {
-            newSize = CGSize(width: size.width * heightRatio, height: size.height * heightRatio)
-        } else {
-            newSize = CGSize(width: size.width * widthRatio, height: size.height * widthRatio)
-        }
-
-        // This is the rect that we've calculated out and this is what is actually used below
-        let rect = CGRect(x: 0, y: 0, width: newSize.width, height: newSize.height)
-
-        // Actually do the resizing to the rect using the ImageContext stuff
-//        DispatchQueue.global().async {
-        UIGraphicsBeginImageContextWithOptions(newSize, false, 1.0)
-        image.draw(in: rect)
-        let newImage = UIGraphicsGetImageFromCurrentImageContext()
-        UIGraphicsEndImageContext()
-
-        return newImage!
-//        }
-    }
-    
-    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
-        self.view.endEditing(true)
-        return false
     }
 }
